For brand entities added req and resp dtos , implementing cqrs with mediatr pattern also brand business ruless and updated base Emtity classes

+++ src/corePackages/Core.Security/Extensions/ClaimsPrincipalExtensions.cs
@@ -15,8 +15,8 @@ public static class ClaimsPrincipalExtensions
         return claimsPrincipal?.Claims(ClaimTypes.Role);
     }
 
-    public static int GetUserId(this ClaimsPrincipal claimsPrincipal)
+    public static long GetUserId(this ClaimsPrincipal claimsPrincipal)
     {
-        return Convert.ToInt32(claimsPrincipal?.Claims(ClaimTypes.NameIdentifier)?.FirstOrDefault());
+        return Convert.ToInt64(claimsPrincipal?.Claims(ClaimTypes.NameIdentifier)?.FirstOrDefault());
     }
 }
